requirements.txt
argh==0.26.2
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
backpack==0.1
bcrypt==3.1.4
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cleo==0.6.8
cryptography==2.3
Faker==0.7.18
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
Jinja2==2.10
lazy-object-proxy==1.3.1
libsass==0.15.1
markdown2==2.3.6
MarkupSafe==1.0
masonite==2.0.30
masonite-cli==2.0.12
masonite-entry==0.1.6
masonite-scheduler==1.0.0
more-itertools==4.3.0
mysqlclient==1.3.13
orator==0.9.7
passlib==1.7.1
pastel==0.1.0
pathtools==0.1.2
pendulum==1.5.1
pluggy==0.6.0
psutil==5.4.6
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pyaml==16.12.2
pycparser==2.19
Pygments==2.2.0
pylev==1.3.0
pytest==3.6.0
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.5
pytzdata==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
simplejson==3.16.0
six==1.11.0
tabulate==0.8.2
tldextract==2.2.0
tzlocal==1.5.1
urllib3==1.23
validator.py==1.2.5
waitress==1.1.0
watchdog==0.8.3
whitenoise==3.3.1
wrapt==1.10.11
